Add VolEsti integrator for approximate integration
